restructure activity detail component
